requirements.txt
pillow==10.1.0
numpy==1.24.3
tensorflow==2.14.0
tensorflow-metal==1.1.0